requirements.txt
aniso8601==9.0.1
asgiref==3.8.1
crispy-bootstrap4==2024.1
Django==5.0.6
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-filter==24.2
django-simple-history==3.5.0
graphene==3.3
graphene-django==3.2.1
graphql-core==3.2.3
graphql-relay==3.2.0
promise==2.3
six==1.16.0
sqlparse==0.5.0
text-unidecode==1.3
